chore: added setBurger basic logic, logic from constructor-container to app remove constructor-container component
