+++ src/components/app/app.module.css
@@ -7,3 +7,14 @@
   justify-content: flex-start;
   align-items: center;
 }
+
+.main {
+  width: 1240px;
+  height: 912px;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 40px;
+  position: relative;
+}
