+++ src/components/burger-ingredients/burger-ingredients.js
@@ -1,10 +1,10 @@
-import PropTypes from 'prop-types';
-import Ingredient from './ingredient/ingredient';
-import TabBar from './tab-bar/tab-bar';
-import Collection from './collection/collection';
-import styles from './burger-ingredients.module.css';
-import { ingredientPropTypes } from '../../../utils/prop-types';
-import { BUN, SAUCE, MAIN } from '../constructor-container';
+import PropTypes from "prop-types";
+import Ingredient from "./ingredient/ingredient";
+import TabBar from "./tab-bar/tab-bar";
+import Collection from "./collection/collection";
+import styles from "./burger-ingredients.module.css";
+import { ingredientPropTypes } from "../../utils/prop-types";
+import { BUN, SAUCE, MAIN } from "../app/app";
 
 const BurgerIngredients = ({
   handleTabSelect,
@@ -38,14 +38,14 @@ const BurgerIngredients = ({
     return { left, right };
   };
 
-  const bun = splitterToColumn(data, 'bun');
-  const sauce = splitterToColumn(data, 'sauce');
-  const main = splitterToColumn(data, 'main');
+  const bun = splitterToColumn(data, "bun");
+  const sauce = splitterToColumn(data, "sauce");
+  const main = splitterToColumn(data, "main");
 
   return (
     <div className={styles.container}>
-      <div className='pb-5 pt-10'>
-        <span className='text text_type_main-large'>Соберите бургер</span>
+      <div className="pb-5 pt-10">
+        <span className="text text_type_main-large">Соберите бургер</span>
       </div>
       <TabBar current={current} onClick={handleTabSelect} />
       <div className={styles.scrollbar}>
