+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,5 +1,5 @@
 .container {
-  height: 912px;
+  height: 100%;
   width: 600px;
 }
