+++ src/components/burger-ingredients/ingredient/ingredient.js
@@ -1,10 +1,10 @@
-import PropTypes from 'prop-types';
+import PropTypes from "prop-types";
 import {
   CurrencyIcon,
   Counter,
-} from '@ya.praktikum/react-developer-burger-ui-components';
-import { ingredientPropTypes } from '../../../../utils/prop-types';
-import styles from './ingredient.module.css';
+} from "@ya.praktikum/react-developer-burger-ui-components";
+import { ingredientPropTypes } from "../../../utils/prop-types";
+import styles from "./ingredient.module.css";
 
 const Ingredient = ({ data, handleOpenModal }) => {
   const { name, image_large, price } = data;
@@ -12,15 +12,15 @@ const Ingredient = ({ data, handleOpenModal }) => {
     <div className={styles.relative} onClick={handleOpenModal}>
       <Counter
         count={1}
-        size='default'
+        size="default"
         className={styles.counter}
-        extraClass='m-1'
+        extraClass="m-1"
       />
       <div className={`${styles.container} pb-10`}>
         <img src={image_large} width={240} height={120} alt={name} />
         <div className={`${styles.priceContainer} mb-2 mt-2`}>
-          <span className='text text_type_digits-default'>{price}</span>
-          <CurrencyIcon type='primary' />
+          <span className="text text_type_digits-default">{price}</span>
+          <CurrencyIcon type="primary" />
         </div>
         <div className={styles.nameConatiner}>
           <p className={`${styles.text} "text text_type_main-default`}>
