--- src/components/constructor-container/constructor-container.module.css
@@ -1,10 +1,0 @@
-.main {
-  width: 1240px;
-  height: 912px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 40px;
-  position: relative;
-}
